Module Knowledge.Documentation

Online Knowledge documentation.

Provides information about declared knowledge entities, such as classes, properties, and agents.

module type Element = sig ... end

A documentation element.

module Agent : Element

Describes Knowledge agents.

module Class : Element

Describes Knowledge classes.

module Property : Element

Describes Class properties.

module Rule : sig ... end

val agents : unit -> Agent.t list

agents () is the list of currently registered agents.

val classes : unit -> (Class.t * Property.t list) list

classes () public classes and their properties.

Only classes and properties that are declared public are included in the list.

val rules : unit -> Rule.t list